removed duplicate function in convert&printf_utils

+++ include/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: mweverli <mweverli@codam.nl>                 +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/06/02 16:15:30 by mweverli      #+#    #+#                 */
-/*   Updated: 2022/08/01 19:13:23 by mweverli      ########   odam.nl         */
+/*   Updated: 2022/11/02 15:25:05 by mweverli      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -55,7 +55,7 @@ int				put_bin_32(va_list list);
 int				put_prc(va_list list);
 
 //				FT_PRINTF_UTILS_01
-char			*ft_itoa(long n);
+char			*ft_itoa(int n);
 char			*ft_itoh(unsigned long n);
 int				write_till(const char **str);
 void			ft_strtoupper(char *str);
